refactor(release): split desktop doctor evaluator
